fix: remove gradio dependency
